fe_configuration: top10features done
