+++ tsfel_example.py
@@ -3,6 +3,12 @@ import numpy as np
 from sklearn.pipeline import Pipeline
 from transformers import FeatureSelectionWrapper, ImputationWrapper, TSFELLagFeatureExtractor
 import time
+from FE_configuration import feature_extraction_preprocessing
+from utils import train_test_split_by_time
+
+
+
+
 
 # Load dataset
 print("Loading data...")
@@ -75,15 +81,23 @@ synthetic_data = {
 
 raw_df = pd.DataFrame(synthetic_data)
 '''
-
-
-# Define column configuration
 COLUMN_CONFIG = {
     'primary_key': ['ProcessId', 'DateTime'],
     'time_col': 'DateTime',
     'target_col': 'event',
     'protected_cols': ['ProcessId', 'DateTime', 'event']
 }
+
+train_df, test_df = train_test_split_by_time(raw_df, time_col='DateTime', id_col='ProcessId', train_ratio=0.7)
+X_train, y_train = train_df.drop(columns=['event']), train_df['event']
+X_test, y_test = test_df.drop(columns=['event']), test_df['event']
+
+feat_extract_config = feature_extraction_preprocessing(train_df, COLUMN_CONFIG)
+
+
+'''
+# Define column configuration
+
 
 # Print initial data info
 print("\nInitial Data:")
@@ -128,3 +142,4 @@ feature_cols = [col for col in transformed_df.columns
                 if col not in COLUMN_CONFIG['protected_cols']]
 print("\nFirst 3 rows with 5 features:")
 print(transformed_df.head(3))
+'''
